DMP-1737 ARM - Pull Response files from ARM
Commiting changes made so far
