+++ src/main/java/uk/gov/hmcts/darts/arm/util/files/UploadFileFilenameProcessor.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.darts.arm.component.files;
+package uk.gov.hmcts.darts.arm.util.files;
 
 import lombok.Getter;
 import lombok.extern.slf4j.Slf4j;
@@ -10,28 +10,28 @@ import static uk.gov.hmcts.darts.arm.service.impl.ArmResponseFilesProcessorImpl.
 
 @Getter
 @Slf4j
-public class CreateRecordFilenameProcessor {
+public class UploadFileFilenameProcessor {
     private static final int NUMBER_OF_TOKENS = 4;
-    private String createRecordFilename;
+    private String uploadFileFilename;
     private String hashcode;
     private String hashcode2; //TODO find out what this field represents
     private String status;
 
-    public CreateRecordFilenameProcessor(String createRecordFilename) {
-        this.createRecordFilename = createRecordFilename;
+    public UploadFileFilenameProcessor(String uploadFileFilename) {
+        this.uploadFileFilename = uploadFileFilename;
         processFilename();
     }
 
     private void processFilename() {
-        List<String> tokens = Arrays.asList(createRecordFilename.split(ARM_FILENAME_SEPARATOR));
-        // CR - Create Record - This is the create record file which represents record creation in ARM.
-        // 6a374f19a9ce7dc9cc480ea8d4eca0fb_a17b9015-e6ad-77c5-8d1e-13259aae1895_1_cr.rsp
+        List<String> tokens = Arrays.asList(uploadFileFilename.split(ARM_FILENAME_SEPARATOR));
+        // UF - Upload File - This is the Upload file which represents the File which is ingested by ARM.
+        // 6a374f19a9ce7dc9cc480ea8d4eca0fb_04e6bc3b-952a-79b6-8362-13259aae1895_1_uf.rsp
         if (tokens.size() == NUMBER_OF_TOKENS) {
             hashcode = tokens.get(0);
             hashcode2 = tokens.get(1);
             status = tokens.get(2);
         } else {
-            log.error("Expected {} tokens in filename {} but found {}", NUMBER_OF_TOKENS, createRecordFilename, tokens.size());
+            log.error("Expected {} tokens in filename {} but found {}", NUMBER_OF_TOKENS, uploadFileFilename, tokens.size());
             throw new IllegalArgumentException();
         }
     }
